requirements.txt
pymysql==1.1.0
numpy==1.24.3
pandas==2.0.3
schedule==1.2.0
flask==2.3.3
SQLAlchemy==2.0.23
matplotlib==3.7.2
scikit-learn==1.3.0
apscheduler==3.10.4
pyyaml==6.0.1
